+++ apps/www/src/external/AgeResult.d.ts
@@ -1,0 +1,5 @@
+export type AgeResult = Readonly<{
+    ages: Readonly<[number, number]>
+    source: string
+    sourceTitle: string
+}>
